Refactor: Split operations.rs into read_ops.rs and write_ops.rs
This commit refactors the `service/src/operations.rs` file by splitting its responsibilities into two new files:
- `service/src/read_ops.rs`: Contains all functions related to reading data (e.g., `read_node_score`, `read_graph`).
- `service/src/write_ops.rs`: Contains all functions related to modifying data or state (e.g., `write_put_edge`, `write_log_level`).

The A* search helper functions and related `AStarError` enum, previously in `operations.rs`, have been moved into `read_ops.rs` as private helper utilities, since `read_graph` was their primary consumer.

Additionally, long functions within both `read_ops.rs` and `write_ops.rs` have been refactored by breaking them down into smaller, private helper functions. This was done to improve readability and ensure that individual functions are more focused and generally fit within a single screen.

The main goals of this refactoring were:
- Improve code organization and separation of concerns (read vs. write).
- Enhance readability and maintainability of the operational logic.
- Reduce the size and complexity of individual functions.

The `operations.rs` file has been removed, and all call sites (primarily in `request_handler.rs`) have been updated to use the new modules. The `NeighborDirection` enum and `neighbor_dir_from` function were added to `protocol.rs` to resolve a compilation issue that arose from the refactoring. Several other minor compilation and borrow-checking issues were also addressed during the process.

All existing tests pass with these changes.
