+++ service/src/main.rs
@@ -4,7 +4,8 @@ pub mod bloom_filter;
 pub mod constants;
 pub mod log;
 pub mod nodes;
-pub mod operations;
+pub mod read_ops;
+pub mod write_ops;
 pub mod protocol;
 pub mod quantiles;
 pub mod request_handler;
